backend/discord_bot/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pydantic>=2.0.0
loguru>=0.7.0 